shopglobal_bq_loader/requirements.txt
google-cloud-bigquery>=3.11.4
pandas>=2.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0
tqdm>=4.65.0
pandas-gbq>=0.26.1 
google-auth
google-cloud-discoveryengine
pydrive2